+++ src/SALOMEDSImpl/SALOMEDSImpl_AttributeTableOfInteger.cxx
@@ -232,9 +232,9 @@ void SALOMEDSImpl_AttributeTableOfInteger::SetColumnData(const int theColumn,
     myTable[myNbColumns*((int)i-1)+theColumn] = theData[i-1]; //!< TODO: conversion from size_t to int
   }
 
-  if(aLength > myNbRows) {
+  if((int)aLength > myNbRows) {
     myNbRows = (int)aLength; //!< TODO: conversion from size_t to int
-    while (myRows.size() < myNbRows) { // append empty row titles
+    while ((int)myRows.size() < myNbRows) { // append empty row titles
       myRows.push_back(std::string(""));
     }
   }
@@ -263,7 +263,7 @@ void SALOMEDSImpl_AttributeTableOfInteger::SetColumnTitle(const int theColumn,
 {
   CheckLocked();                                                      
   Backup();
-  while(myCols.size() < theColumn) myCols.push_back(std::string(""));
+  while((int)myCols.size() < theColumn) myCols.push_back(std::string(""));
   myCols[theColumn-1] = theTitle;
   
   SetModifyFlag(); //SRN: Mark the study as being modified, so it could be saved 
@@ -272,13 +272,13 @@ void SALOMEDSImpl_AttributeTableOfInteger::SetColumnTitle(const int theColumn,
 std::string SALOMEDSImpl_AttributeTableOfInteger::GetColumnTitle(const int theColumn) const 
 {
   if(myCols.empty()) return "";
-  if(myCols.size() < theColumn) return "";
+  if((int)myCols.size() < theColumn) return "";
   return myCols[theColumn-1];
 }
 
 void SALOMEDSImpl_AttributeTableOfInteger::SetColumnTitles(const std::vector<std::string>& theTitles)
 {
-  if (theTitles.size() != myNbColumns) throw DFexception("Invalid number of columns");
+  if ((int)theTitles.size() != myNbColumns) throw DFexception("Invalid number of columns");
   size_t aLength = theTitles.size(), i;
   for(i = 0; i < aLength; i++)  myCols[i] = theTitles[i];
   
@@ -315,7 +315,7 @@ void SALOMEDSImpl_AttributeTableOfInteger::PutValue(const int theValue,
   myTable[anIndex] = theValue;
 
   if(theRow > myNbRows) {
-    while (myRows.size() < theRow) { // append empty row titles
+    while ((int)myRows.size() < theRow) { // append empty row titles
       myRows.push_back(std::string(""));
     }
     myNbRows = theRow;
@@ -442,11 +442,11 @@ std::string SALOMEDSImpl_AttributeTableOfInteger::Save()
 {
   std::string aString;
   char* buffer = new char[1024];
-  size_t i, j, l;
+  int i, j, l;
 
   //Title
-  l = myTitle.size();
-  sprintf(buffer, "%u\n", l);
+  l = (int)myTitle.size();
+  sprintf(buffer, "%d\n", l);
   aString+=buffer;
   for(i=0; i<l; i++) {
     aString += myTitle[i];
@@ -459,8 +459,8 @@ std::string SALOMEDSImpl_AttributeTableOfInteger::Save()
 
   //Row titles
   for(i=0; i<myNbRows; i++) {
-    l = myRows[i].size();
-    sprintf(buffer, "%u\n", l);
+    l = (int)myRows[i].size();
+    sprintf(buffer, "%d\n", l);
     aString+=buffer;
     for(j=0; j<l; j++) {
       aString += myRows[i][j];
@@ -474,8 +474,8 @@ std::string SALOMEDSImpl_AttributeTableOfInteger::Save()
 
   //Columns titles
   for(i=0; i<myNbColumns; i++) {
-    l = myCols[i].size();
-    sprintf(buffer, "%u\n", l);
+    l = (int)myCols[i].size();
+    sprintf(buffer, "%d\n", l);
     aString+=buffer;
     for(j=0; j<l; j++) {
       aString += myCols[i][j];
@@ -484,8 +484,8 @@ std::string SALOMEDSImpl_AttributeTableOfInteger::Save()
   }
 
   //Store the table values
-  l = myTable.size();
-  sprintf(buffer, "%u\n", l);
+  l = (int)myTable.size();
+  sprintf(buffer, "%d\n", l);
   aString+=buffer;
   for(MI p = myTable.begin(); p != myTable.end(); p++) {
     sprintf(buffer, "%d\n%d\n", p->first, p->second);
